Createtd new functionality
1. Relationships
2. Add Ownership to Listings
3. Manage Listings Page
4. User Authorization
And I used the command tinker

+++ app/Http/Controllers/ListingController.php
@@ -50,6 +50,8 @@ class ListingController extends Controller
             'public');
         }
 
+        $formFields['user_id'] = auth()->id(); 
+
         Listing::create($formFields);
 
 
@@ -64,6 +66,11 @@ class ListingController extends Controller
 
      //Update Listig Data
      public function update(Request $request, Listing $listing) {
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -86,8 +93,19 @@ class ListingController extends Controller
     }
 
     //Delete Listing
-    public function destroy(Listing $listing){
+    public function destroy(Listing $listing) {
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
-        return redirect('/')->with('message', 'Listing deleted Succesfully');
+        return redirect('/')->with('message', 'Listing deleted succesfully');
+    }
+
+    //Manage Listing
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()
+        ->user()->listings()->get()]);
     }
 }
